Move ten Models files into Text-Grab.Core and finish the FileOpenUtilities split
Batch 2b. All ten files read in full and verified clean per invariant 8:
AsyncOcrFileResult, EditTextTableDocument (moved first - FindResult calls
its static GetSpreadsheetColumnLabel), ExtractedPattern (uses
StringMethods.ExtractSimplePattern, already in Core), FindResult,
GrabFrameTableEditState, GrabFrameWordGroupingMode, SpreadsheetUndoHistory
(internal - already covered by Core's existing InternalsVisibleTo("Tests")
and ("Text-Grab")), TemplatePatternMatch, TemplateRecognizerMatch (uses
RecognizerOutputKind, already in Core), ThirdPartyPackageInfo. No call
sites needed fixing - everything resolved through the existing
Text_Grab.Models namespace.

This unblocks the other half of the FileOpenUtilities revisit deferred
in the previous commit: EtwEditorMode moved to Core along with
EditTextTableDocument.cs, so GetEditorModeForPath now moves too, joining
GetOpenContentKindForPath in Text-Grab.Core/Utilities/IoUtilities.cs.
FileOpenUtilities.cs is left with only GetContentFromPath and
TryToOpenTextFile, which stay app-side (OcrUtilities call, Wpf.Ui/WinForms
MessageBoxes). Updated call sites: EditTextWindow.xaml.cs and
Tests/FilesIoTests.cs.

Gate: Tests.Core 14/14, Tests.Core.Windows 3/3, Tests 1493/7 skipped/0
failed - matches baseline.
