+++ Text-Grab.Core/Models/AsyncOcrFileResult.cs
