+++ Text-Grab.Core/Models/EditTextTableDocument.cs
