+++ Text-Grab.Core/Models/ExtractedPattern.cs
