+++ Text-Grab.Core/Models/FindResult.cs
