+++ Text-Grab.Core/Models/GrabFrameTableEditState.cs
