+++ Text-Grab.Core/Models/GrabFrameWordGroupingMode.cs
