+++ Text-Grab.Core/Models/SpreadsheetUndoHistory.cs
