+++ Text-Grab.Core/Models/TemplatePatternMatch.cs
