+++ Text-Grab.Core/Models/TemplateRecognizerMatch.cs
