+++ Text-Grab.Core/Models/ThirdPartyPackageInfo.cs
